+++ Contents/mods/ChaosMod/42/media/lua/client/ChaosMod/NPC/ChaosNPC.lua
@@ -58,6 +58,10 @@ require "ChaosMod/NPC/ChaosNPCConstants"
 ---@field maxHealth number
 ---@field chanceToDropWeaponOnDeath number
 ---@field lastNeedHealLineMs integer
+---@field panicCheckTimeoutMs integer
+---@field panicStartTimeMs integer
+---@field panicTargetSquare? IsoGridSquare
+---@field panicCooldownEndMs integer
 ChaosNPC = ChaosNPC or {}
 ChaosNPC.__index = ChaosNPC
 ChaosNPC._nextGroundWeaponClaimId = ChaosNPC._nextGroundWeaponClaimId or 0
@@ -122,6 +126,10 @@ function ChaosNPC:new(zombie, nickname)
     o.maxHealth = 1.0
     o.chanceToDropWeaponOnDeath = 0.4
     o.lastNeedHealLineMs = 0
+    o.panicCheckTimeoutMs = 0
+    o.panicStartTimeMs = 0
+    o.panicTargetSquare = nil
+    o.panicCooldownEndMs = 0
     ChaosNPC._nextGroundWeaponClaimId = ChaosNPC._nextGroundWeaponClaimId + 1
     o.actionWorldObjectClaimToken = "npc_ground_weapon_claim_" .. tostring(ChaosNPC._nextGroundWeaponClaimId)
     return o